Use more generic apply_filters in jenkins source
Move the checks to utils/filtering.py so that they can be reused by
other sources if necessary and use apply_filters to apply the checks,
simplifyng the code in the Jenkins source.
